web-animations/animation-player/animation-player-playback-rate.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player playbackRate attribute test</title>
<meta name="assert" content="Returns the playback rate of this player. Setting this attribute follows the procedure defined in section 3.5.3.1 Updating the playback rate.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-playbackRate">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function newAnimationPlayer(animationTarget) {
    var animation = new Animation(animationTarget, [], 5);
    return document.timeline.play(animation);
}


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    assert_not_equals(player.playbackRate, null,
        'The value of AnimationPlayer.playbackRate attribute should not be null');
    assert_equals(player.playbackRate, 1,
        'The initial value of AnimationPlayer.playbackRate attribute should not be 1');
}, 'Test initial value of AnimationPlayer.playbackRate attribute');


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    var newPlaybackRate = player.playbackRate + 1;
    player.playbackRate = newPlaybackRate;

    assert_equals(player.playbackRate, newPlaybackRate,
        'The value of AnimationPlayer.playbackRate attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.playbackRate attribute');


test(function() {
    var player = newAnimationPlayer(createDiv(this));

    var newPlaybackRate = -1;
    var previousTime = player.currentTime;
    player.playbackRate = newPlaybackRate;

    assert_equals(player.playbackRate, newPlaybackRate,
        'The value of AnimationPlayer.playbackRate attribute should be set to the new value');
    assert_equals(player.currentTime, previousTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the ' +
        'effective current time before setting new playback value');
    assert_false(player.finished, 'Finished flag should be set to the false')
}, 'Test setting value of AnimationPlayer.playbackRate attribute. ' +
    'New value for the playback rate differs in sign');
</script>
</body>
